+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 ## [0.1.3] - 2017-11-11
 ### Added
+- `clustering:` Ported `KMeans` `fastSquaredDistance()`, `findClosest()` and `pointCost()`
 - `linalg:` Ported `BLAS` `axpy()`, `dot()`, and `scal()`
 - `linalg:` Ported `Vectors` `sqdist()`
 - `linalg:` Implement `VectorWithNorm` equality (`__eq`)
 - `util:` Ported `MLUtils` `EPSILON` and `fastSquaredDistance()`
 
 ### Changed
+- `clustering:` Fixed KMeans getters and setters to match RDD API, not DataFrame API
 - `linalg:` Vector types are now 0-based like Apache Spark
 - Consolidate Apache Spark and Stuart unit tests into a single unified folder hierarchy